requirements.txt
boto3==1.16.43
django==3.1.4
django-environ==0.4.5
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
psycopg2-binary==2.8.6
